Updated openstack/openstack Project: openstack/python-novaclient  a820c45543dc843c1b75aa65903558a51b4ba037
Use v2.0 in clouds.yaml auth_url for functional test job

Commit 050a0d5b304a013e23cd5909abf6e11b7dda5f18 changed the defaults
in devstack to use the keystone v3 api. This broke the novaclient
functional test job because novaclient doesn't yet support v3
and relies on having "v2.0" in the auth_url.

This appends or replaces the version in the clouds.yaml auth_url
to v2.0 to enable the functional test job to work until v3 support
is added to novaclient.

Closes-Bug: #1506103

Change-Id: I75c5efd039b71c8855acf21c5516ccbeeaaeadb3