Updated openstack/openstack Project: openstack/horizon  7bcb859ee74157ff0bcde4af7bef32453585c185
Prevent the click loss at navbar accordion in integration tests

Simplify code for handling the navbar in integration tests, replace
the loop with explicit waits for the menu items to receive the proper
classes (which mark collapsed/expanded/in-transit states). Besides
fixing intermittent failures this also clarifies the overall logic.

Change-Id: I7a58a59be66264ce6e3c05f78033961954f0c553
Closes-Bug: #1505320