Updated openstack/openstack Project: openstack/neutron  74ed514596993c06b94a3b9154acbc32337334b8
Do not try to delete a veth from a nonexistent namespace

Currently, VethFixture cleanup tries to delete veth from their
namespaces even if they don't exist anymore (without crashing the
cleanup). It implies an extra trace if an error is raised which can be
confusing.

This change avoids to try deleting a veth from a nonexistent namespace
in the fixture VethFixture.

Closes-Bug: #1506862
Change-Id: I5d0192998d2f3f8d1a6f783769ae9bfcb4bae7f2