Updated openstack/openstack Project: openstack/horizon  1b0fe09e3d1fecf8113f05b423be5e4c2d091128
admin not able to set maintenance state for volume

In admin volume table, maintenance state was not able
to set by admin via horizon. since the option is not
present in the choice list.

The status is added in the display choices also to
show the translated status.

This patch adding the new maintenance state in the
choice list.

Change-Id: I11c3845fb95093db08c5a0fdb568dd06b7370841
Closes-Bug: #1490559