Updated openstack/openstack Project: openstack-dev/bashate  5c7a3a125a308f0bd4beaa8c943a952383d38b29
Replace heredoc fail-to-end detection E012 with "bash -n"

Currently E012 (detect if heredoc doesn't end) basically doesn't work;
it relies on having multiple files specified to check.  The logic is
currently that if a heredoc is open and we start a new-file, we missed
it.

As we even say in the comments, bash will warn about this.  Luckily,
"bash -n" does pop out this warning when we forget to close a
heredoc.

So modify the existing bash syntax check to catch this too.  Because
it's a warning, bash doesn't exit with !0, so rework things to just
parse output in all cases.  I also found a bit of inconsistency with
output, as described in the comments.

Remove the existing check, and add a testcase.

Change-Id: I292a926d3297a283f74f7d47b7c7aa3c42c9f030