Updated openstack/openstack Project: openstack/requirements  9cb2ca8acc558a5a3e676a4d72be429ac32c9cc3
global-requirements.txt: Bump min. version of oslo.reports

There was a recent code change[1] in oslo.reports project that modified
what signal is expected to correctly generate Guru Meditation Error
reports.  This change was included in the just released[2] Oslo 0.6.0
version.

This version bump is a prerequisite for the corresponding Nova
documentation patch[3], which will align the documentation instructions
with the said code change in oslo.reports.

[Thanks to Matt Riedemann for raising this point.]

[1] https://review.openstack.org/#/c/223133 --
    guru_meditation_report: Use SIGUSR2 instead of SIGUSR1
[2] https://review.openstack.org/#/c/233326/ --
    Oslo Releases for Week starting Oct 12th
[3] https://review.openstack.org/#/c/227779 --
    doc: gmr: Update instructions to generate GMR error reports

Change-Id: Idaf8fcaee752fbd645246982c4480521cc2e659a