Updated openstack/openstack Project: openstack/nova  5ee1eac516e428b5f71dfcc34cdfc71c484a789b
Remove redundant deps in tox.ini.

Tox installs the current tree (either via an sdist or as a develop
install) always. Our requirements.txt are reflected into the metadata
of that tree, so we shouldn't use requirements.txt directly - it just
means that we'd be hiding from ourselves issues like e.g. using URLs
which can't be reflected properly.

Change-Id: Id096f1d1971cb8a19f9e86ba57957b39e00bd6f5