Updated openstack/openstack Project: openstack/neutron  60b957fb72228d3325fd05a312372e23ba892eb8
Revert "DVR: Notify specific agent when update floatingip"

This reverts commit 52e91f48f2327b47f126893f9cb12f153380a9a6.

The patch broke notifications about FIP updates and triggered 100%
gate failures for Ironic gate.

I believe that I0cbe8c51c3714e6cbdc48ca37135b783f8014905 is also
breaking notifications, but for FIP create, which probably was not
utilized in any gate before and hence not caught in time.

The change the reverted patch introduced made update_floatingip to
fetch router based on FIP router_id field on every call, which was
not the case before the patch. For some reason unknown at the
moment, we get NotFound from database on this fetch.

The patch does not answer the question why we get NotFound from
database on fetching a FIP router_id, but that's another issue that
should be investigated while Ironic gate is happy.

Change-Id: I4affac49d7c63f47c5654b94b28f4cb7471e87b0
Closes-Bug: #1507558
Related-Bug: #1507602