Updated openstack/openstack Project: openstack/keystone  a86d979146a12c581a8cd0ca7672a888db493fa6
Fixes kvs cache key mangling issue for Py3

The dogpile sha1_key_mangler function was not safe to use with text
data. A new default key mangler was added base on the one that is
included with oslo_cache.  All of the test modules that were blocked
by this have been enabled.

bp python3

Change-Id: Id651b01a0def8c4043eebb56d7e01d62ff4251f8