Updated openstack/openstack Project: openstack-infra/system-config  93a1846b36e427996b6913fa064b13c48792be16
Add openstack-health api server

This commit adds the puppet config for running an openstack-health
api instance. It'll use the subunit2sql trove db node as a data
source. The puppet-openstack-health module is incorrectly named which
is causing issues, a workaround to rename the repo is added to
install_modules.sh. This is a temporary measure until the gerrit
rename is completed. There is also a workaround in the apply test.

Change-Id: I7e6d9664d087e7bdc21d92624991d0d5f86c0c99