Updated openstack/openstack Project: openstack/oslo.config  ca1ca6a3a90edb0bbe862b2a77e6344fbdcb4f75
Raise exception when wrong type is used as default

In I12cd5f4be82ac5ed0ebb348b358dd8eaf04e4e0d the change introduced type
checking of the default value to the Opt and logged a warning. This
change now raises an exception as per the comment in the original
change.

Closes-Bug: #1502072
Change-Id: I9b806efc750009d64b08a5583449bebbf10e3fb4