Updated openstack/openstack Project: openstack/cinder  60fac91e87b0618f506fc96cd3b6b9f8a42ea8e6
Ensure replication functions check driver status

This patch adds the missing checks to ensure
the driver's are initialized to both
enable_replication and disable_replication.

This patch also adds comment blocks to the
failover_replication as well as list_replication_targets
functions to explain why they are both are not
checking the driver for initialization.  One of the main
use cases for those 2 functions happen precisely when the
driver can't talk to the primary backend, but might be
able to talk to the secondary or target backends.

Change-Id: I4052e0ff4d2977ecfba940db56dba24c063bbbb6
Closes-Bug: #1501936