Updated openstack/openstack Project: openstack-infra/project-config  e4e170b3e638accd2ad69746cd8d4fbdfe14ffd3
Make openstack-health npm lint job voting

We have finally gotten the lint checking into a good spot on
opnestack-health, so let's make the job voting so we're enforcing this
on new commits. It'll also validate that we can actually build the js
being submitted, which is something we're missing right now.

Depends-On: I3aa0c91e77f40c6383ac73850f5b93ebae9af323

Change-Id: Ie3d7326b8a3cc7f01d7ded174a0ac077e9695054