Updated openstack/openstack Project: openstack/manila  099145a07bdbc15b332d59c840d155de37a819d9
Merge tag '1.0.0'

This is a null-merge of the 1.0.0 release tag back into the master
branch so that the 1.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I9f6bd68468f837c1f28e1428ef10d1562a5c1340