Updated openstack/openstack Project: openstack-infra/project-config  9a969c9ba1222242c4a638b2521c3fa5525d14f6
Add openstack-resource-agents project

The Pacemaker Resource Agents for OpenStack project is a revitalization
of a long-used unofficial repository which is referenced by the official
OpenStack High Availability guide but no longer maintained:

  https://github.com/madkiss/openstack-resource-agents

All relevant stakeholders (including the original maintainer) have been
consulted and given unanimous support:

  https://github.com/madkiss/openstack-resource-agents/issues/22
  http://lists.openstack.org/pipermail/openstack-dev/2015-June/067763.html
  http://clusterlabs.org/pipermail/users/2015-June/000732.html
  http://oss.clusterlabs.org/pipermail/pacemaker/2015-June/023649.html

The import is actually from this fork of the origin repository:

  https://github.com/aspiers/openstack-resource-agents

in order to allow replacement of the old stable branches with tags prior
to import.  There are no other differences between the fork and the
original.

New launchpad project is at:

  https://launchpad.net/openstack-resource-agents

Later we'll review the CLA policy, set up some CI (currently there is no
test suite) etc., but for now this is a first step to import the code
and set up the corresponding #openstack-ha IRC channel.

Signed-off-by: Adam Spiers <aspiers@suse.com>

Change-Id: I8917ab2ec3b02d26de0a1881eeb676e63e5725d4