Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  dbe63dd9d209eabfcb7c562c17b379abaf79a0dc
cmd: Add a config file fallback to the user's home directory

The system-global path of

    /etc/jenkins_jobs/jenkins_jobs.ini

is not applicable on Windows. To solve that, and as Linux users would
benefit from a per-user configuration file, too, introduce a fallback at

    ~/.config/jenkins_jobs/jenkins_jobs.ini

following the XDG Base Directory Specification. Try to use that per-user
file if no configuration file is given on the command line and none is
present in the script directory for development purposes.

Change-Id: I1a22454999ee8ac60d6f22545850c33aa03bb29b