Updated openstack/openstack Project: openstack/tripleo-heat-templates  88efef101e3ce58d2c9e3f53070df936a497b518
Allow ctlplane network or ID, and default to "ctlplane"

Currently you always have to pass the ctlplane ID because we're still
using the deprecated network_id property for the neutron port resource.

Since Juno, heat has supported a "network" property, which is used
elsewhere, e.g the nested port stacks, so switch to using it in the
overcloud-without-mergepy template, and flip the default to a more
useful "ctlplane" vs an empty string.

This means the stack create should just work on commonly documented
deployments without requiring any parameter.

Change-Id: Ifcea36d26b795c5e8b80accd8112e23b254127be