Updated openstack/openstack Project: openstack-infra/devstack-gate  4ed122755489bd7f9d127afd8d59e74addf6ca95
Remove jeepyb and pypi-mirror from PROJECTS

These two projects were used as part of the setup for the
requirements/pbr testing but that is no longer the case. In fact
requirements and pbr both have special casing to ignore these projects
in the PROJECTS list now. Remove them to shorten the PROJECTS list which
simplifies test run workspace setup.

Change-Id: I61c2f3f8a724dc789479278e7ba493e095c52305