Updated openstack/openstack Project: openstack/neutron  07dda9709aec13c949e7953b187c4ab8be3e1227
Properly handle segmentation_id in OVS agent

The segmentation_id of a OVS VLAN can be None, but a recent change
assumed that it was always an integer. It highlighted the fact that we
try to store None in the OVS database, which got stored as a string.
This fixes the storage, and handles loading the value while keeping
compatibility.

Change-Id: I6e7df1406c90ddde254467bb87ff1507a4caaadd
Closes-Bug: #1494281