Updated openstack/openstack Project: openstack/horizon  e62cdfaadb00c5d124d575077ece18269a214379
Change ports range in validators and test function

Current implementation for validating the port number and protocol
number accepts invalid values. For port number it accepts -1 which is
invalid, and the 0 which is used as programming technique for specifying
system-allocated (dynamic) ports. For protocol number it accepts only
one invalid value which is -1.

Proposed solution changes the range of:
- valid port numbers from <-1, 65535> to <1, 65535>
- valid protocol numbers from <-1, 255> to <0, 255>

Co-Authored-By: Wlodzimierz Borkowski <wlodek.borkowski@gmail.com>
Co-Authored-By: Kamil Rykowski <kamil.rykowski@intel.com>

Closes-Bug: 1329571
Closes-Bug: 1431367
Change-Id: I2dc83b611deaa0cc91d586dad1c01d8de91d1957