Updated openstack/openstack Project: openstack-infra/project-config  154bbe2a29dc270d0e427445b61bcb2a033e8a5c
Rename project networking-bagpipe(-l2)

The networking-bagpipe-l2 project needs to be renamed
to reflect the fact that its scope encompasses more than just
a driver for Neutron L2 networks. The scope includes agent
extensions for BGPVPN interconnection, and will extend to Neutron
L3 and service-chaining.

The renaming consists in simply removing "-l2" from the name.

The group statement in gerrit/projects.yaml is not needed anymore,
because the LP project has been removed to drop "-l2" and
add 'networking-" so that LP and Openstack project are now identical.

This change also updates the gerrit ACLs for networking-bagpipe
to match the access list of a neutron stadium project.

Change-Id: I7774b9e9eb830c9320b361b78ebb0dbde3f97205