Updated openstack/openstack Project: openstack/neutron  5cdab68528d165a654bf9b5a9822f586e0d51ee4
Refactor _populate_ports_for_subnets for testability

I want to add to this function but I found it very difficult to test
the additions that I was making (because this code isn't well covered
by UTs).  This patch adds some unit testing so that I can change it
more confidently but does not change any functionality.

Change-Id: Id2cf57968f2dfb8bcf18946c922ce2bb51a27766
Partially-implements:  bp/address-scopes