Updated openstack/openstack Project: openstack-infra/project-config  9f29053016ae2398a58808b848dac84690fe6c4b
Deletegate gerrit project ownership

The openstack-infra/gerrit project[1] tracks upstream Gerrit and also serves
as a fork which we use for review.o.o.  There is a gerrit-core group[2] but
it's not being used yet.  Change ownership of the project from infra-core
to gerrit-core so that we can delegate ownership to a subset of devs working
to maintain openstack's Gerrit.

[1] https://git.openstack.org/cgit/openstack-infra/gerrit
[2] https://review.openstack.org/#/admin/groups/921,members

Change-Id: Id022177dc9ebb8755f334ff02f1aa60e36578056