Updated openstack/openstack Project: openstack-infra/project-config  ccddbedc3ee48a0da25ba77717006cc175773dbc
Enable stable branches for tripleo repositories

Adjust the acls so the following repos allow branch creation,
administered by the tripleo-release group, with reviews of the
to-be-created stable/liberty branches handled by the existing
tripleo-core team (we're not planning a separate stable-maint
group at this point).

The following repos are affected:

* openstack/tripleo-common
* openstack/tripleo-docs
* openstack/tripleo-heat-templates
* openstack/tripleo-puppet-elements
* openstack/python-tripleoclient
* openstack/instack-undercloud

These now all share the common tripleo.config, with instack split
out to a seperate acl, as it does not currently require a stable
branch.

This is the first step to implementing the stable/release branch
process described in https://review.openstack.org/#/c/221811/

Change-Id: Ibd82864cc8d25acf0f8c425d13f1ccdffff50a31