Updated openstack/openstack Project: openstack-infra/devstack-gate  62ea8f2e93d970a0b4dcf1ad66b0d74294d30976
Remove os-loganalyze from PROJECTS

os-loganalyze was added to PROJECTS so that it could be used in
devstack. We don't currently run any jobs that use os-loganalyze with
d-g so we don't actually need this. Any new jobs that do want to use it
can be explicit about requiring it in their PROJECTS list.

Change-Id: I4db4fe1511c32fa16d47cda0a11bf44815646ede