Updated openstack/openstack Project: openstack/python-keystoneclient  574fad7abb328276396d7d4c2743982bb903078c
pass on @abc.abstractmethods

Although a bare docstring is entirely valid grammar, those new to Python find
that pattern baffling, *especially* when other similar methods contain pass
statements. So, for consistency, add a pass statement to otherwise bare
@abc.abstractmethods.

Note that the implementation of an @abc.abstractmethod (in the abstract
base class) can still be called by concrete children, so suddenly
raising a NotImplementedError() instead might be "surprising" to
implementors. A no-op such as "pass" or "return None" is preferable.

Change-Id: I79969ad1a3429516ea785c649a165ead54944225