Updated openstack/openstack Project: openstack/oslo.messaging  87ad4c77ff8e0c6df68e8c1c142415fefc660032
Rename MessageHandlingServer._executor for readability

MessageHandlingServer has both MessageHandlingServer.executor, which
is the name of an executor type, and MessageHandlingServer._executor,
which is an instance of that type. Ideally we would rename
MessageHandlingServer.executor, but as this is referenced from outside
the class we change _executor instead to _executor_obj.

Change-Id: Id69ba7a0729cc66d266327dac2fd4eab50f2814c