Updated openstack/openstack Project: openstack-infra/devstack-gate  be9819e80be46c3a331e0cd8edef84f3f4cfc448
use cp always instead of mv / cp logic

Grenade was not getting cached files on the old side, leading to
higher failure rates as it talks to the internet for things like the
cirros images. Part of this is because there is conditional mv / cp
logic that might have worked once, but doesn't now.

In order to address space concerns, do the following:

mv the cache files to $BASE/cache/files (which is typically going to
be /opt/stack/cache/files), this should be on the same filesystem as
devstack / grenade in all of our cloud providers.

cp -l from this new cache to where the devstack / grenade caches
should be. This means we'll still only have 1 copy of the bits in the
filesystem, now with a number of hardlinks to it.

Closes-Bug: #1507533

Change-Id: Ie09e1671329c4144be85af71ea892246b81fdad3