Updated openstack/openstack Project: openstack-infra/system-config  0981ffebcf9d9ddb3a45a68556431c617f66b8d5
Add openstack-health to status.o.o vhost template

We recently added the openstack-health frontend to be hosted on
status.o.o and are currently building the js and moving it to the
right location. However we neglected to also update the vhost template
to tell apache about the new directory to host. This commit updates
the template to ensure we are properly hosting the frontend.

Change-Id: I17efdde03e41c5d4b563b9eef8cba8c42a8ab439