Updated openstack/openstack Project: openstack-infra/project-config  1b9a935b64b74677a88dea2056f9422a88295233
Check for valid upstream git projects

This checks out the upstream git projects (unless we use track-upstream)
and prints some information about it.

Note this checks for all upstream repositories but since those are
cleaned up regularly, this should be fine.

It fails if the there is no remote master branch.

Reindent the original code as main function.

Change-Id: Ie469905f48e4bc1a408fb9e731f95b2b735a6a02