Updated openstack/openstack Project: openstack/oslo.config  f6c668bfb32cbde63866541e8c2100eda5fa5c66
Add hostname config type

Hostnames are often used in config files and they have specific
validation requirements. This change adds a config to validate a
correct hostname are specified in a config file.

A hostname refers to a valid DNS or hostname. It must not be longer than
253 characters, have a segment greater than 63 characters and start or
end with a hyphen.

Closes-Bug: #1508943
Co-Authored-By: ChangBo Guo(gcb) <eric.guo@easystack.cn>
Change-Id: Ic1028a437aee6076c4b7f437f60bbd209f38a20e