Updated openstack/openstack Project: openstack/neutron  feed0973970f850e97db0c7f7c4fdbeedd9a2753
Minor improvement in port_bound operation

By returning the need for binding, the (no-op) port processing
is skipped in the main sync loop. In fact, the port is already
detected as deleted and avoiding going over it again will save
us one log message, and thus many, many, many, many bytes over
time in logs.openstack.org.

If this is not an improvement, I don't what is.

Related-bug: 1455320

Change-Id: I4ef30bed6a1619f5980c3e61ee23ffb808c2c4f0