Updated openstack/openstack Project: openstack/ironic-python-agent  64bdf16b681df938bfdb09cd9d723085a9f8b059
Refactor backoff looping call

This codebase can now be updated to use the refactored
base class to avoid duplicating alot of the same code.

Eventually we should also just move this to oslo if
when/this review is accepted since that seems like a better
home for this code in general.

Change-Id: I387d60667f427824a64d52544638792429887ebf