Updated openstack/openstack Project: openstack/openstack-manuals  c0cd0f04f0802026e9bc6392de86934f15fe4399
Combine contributor quickstart setup steps 1 and 2

Treat account setup and ICLA as one step in the Contributor Guide
Quickstart chapter. The old second step merely linked to a
subsection of the Infra Manual section linked in the first step. If
it's unclear that the CLA subsection of the Account Setup section is
mandatory, then we should improve the formatting/flow in the Infra
Manual accordingly.

Change-Id: Ie126cbc723fe11cf9b581ea87aa4c9c8f147f3b2