Updated openstack/openstack Project: openstack/ironic  1b3d69f33323e3d2fbdd4986768b0591f63c8a8d
DRAC BIOS vendor_passthru: enable rebooting the node

Changing the BIOS settings require rebooting the node. If the node was
originally powered off, the driver would need to turn it on, poll the
job status on the DRAC card and power off the node. To keep things
simple, the BIOS mgmt vendor_passthru methods avoided doing that.

However, it would be beneficial for the user to let them decide whether
they want to reboot the node automatically after the config job
succeeds. DRAC cards already support this, so the driver doesn't need to
supervise the config job status. With this patch the DRAC vendor
passthru for commit_bios_config accepts a 'reboot' parameter.

Closes-bug: #1508459
Change-Id: Id9a87f6dec6a8d6160948e39a763134729cead99