Updated openstack/openstack Project: openstack-infra/project-config  3a3d50b1201c73cefc2f9064aa54a3ead55b70af
Properly deal with magnum-ui for Zanata

magnum-ui is a special snowflake in that its project name is magnum-ui,
but the directory base for source and translations is magnum_ui. Since
it's a Django project, it also isn't treated like a general Python
project anyway, so treat it like django_openstack_auth, by performing
a little refactoring.

Since the repo now contains files, and is marked for translations in
projects.yaml anyway, enable it for translations jobs.

Change-Id: I7eb73181cccae6d4bfb81848d39099d3502a044c