Updated openstack/openstack Project: openstack/glance  6307e5db14aa413e1fff845da420229f1acd9c5e
Merge tag '11.0.0'

This is a null-merge of the 11.0.0 release tag back into the master
branch so that the 11.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iff3a4a4547842167443e507a739d839e750e1ef9