Updated openstack/openstack Project: openstack/nova  fe51c46ff516b9be08ebc6a54f1cb1000ec667a0
Merge tag '12.0.0'

This is a null-merge of the 12.0.0 release tag back into the master
branch so that the 12.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Ib810278ecb6906aca39a7c18112fa45c1663f1e3