Updated openstack/openstack Project: openstack-infra/system-config  36ac4f52ed22fee6bc4493151240be97c15e6464
Make hiera default values optional

Currently, all hiera calls need to have default values in order for the
apply test to work. This is done by using sed to replace the hiera call
with the default value in the call, i.e. 'hiera("foo", "bar")' is
replaced with '"bar"'.

This adds another sed command that replaces a hiera call that has no
default value specified with the name of the key appended with
NoDefault, i.e. 'hiera("foo")' will become '"fooNoDefault"'.

By making the hiera default values optional, we no longer need to pass
in a default value to every hiera call, and so we can remove the
placeholder XXX that were previously needed by the apply tests but not
by puppet itself.

Change-Id: Ibdeb0c2e49245ea05116e8043936d5b012f46666