Updated openstack/openstack Project: openstack-infra/system-config  4879472db5a57cbe0b8d5247b82af2733f7e1f48
Add echo test to asterisk

This extension will give the user the ability to test there SIP
connection with Asterisk. Basically, when somebody dials 5000, they
will hear a message stating it is an echo test and allow the person to
talk into their phone.  Asterisk will echo back when it hears.

Change-Id: I4b4002ce4ec0a3de3f434754aff9d17d6b82fcb8
Signed-off-by: Paul Belanger <pabelanger@redhat.com>