Updated openstack/openstack Project: openstack-infra/system-config  aefb3c3a04dc73cba199944368a55a4a4f7660cd
Clean up tabs and whitespace in status/index.html

Currently the status/index.html file has lots of extra whitespace
and tabs that it doesn't need. This patch removes them and reformats
the file so that future copy/pasters in this file have a bit better
start to work from.

Change-Id: Idedc3697ce68d7ea976417090499bec7f9c36c8b