Updated openstack/openstack Project: openstack/python-manilaclient  7962533b94ae4abf48b3bdbc65564e7c1d6e6c7f
Fix usage of setUpClass method in functional tests

Class method 'setUpClass' is executed for each test of test suite that
appeared in new thread. It leads to creation of redundant resources,
and, hence, waste of resources.

So, add 'group_regex' option in testr config that will make all tests
of each class run in one thread.

Change-Id: I8100a98961791cee52b0866c7d593cc8371b3329
Closes-Bug: #1509460