Updated openstack/openstack Project: openstack-dev/pbr  934491c0931ccddfa426ecdaf6900b20d6867278
doc: Restructure 'Usage' section

The usage section isn't very clear: it assumes a lot of setuptools
knowledge on behalf of the user when one of the goals is to abstract
away this knowledge.

Begin reworking it by restructuring the existing content. Move the
'setup.py' sub-section to the top, as it's arguably the most important
file, then slightly rework the 'setup.cfg' sub-section.

Change-Id: I85e136e8817e294ce9b536b4f3ee7f262cc3cd2c
Partial-bug: #1509376