Updated openstack/openstack Project: openstack-dev/grenade  a094593dceb8cccb4658335c393439682c0bf006
Namespace XTRACE calls

As a follow-on to Iba7739eada5711d9c269cb4127fa712e9f961695 in
devstack, where we rename the XTRACE storage calls to not conflict, do
the same in grenade

Also some minor turning-down of tracing when importing stackrc; it
floods the logs with irrelevant stuff.  And now we're tracing
properly, cleanup "echo_summary" output in the logs.

Change-Id: Ibb67a6454e4465a8d0493461e32dbae96c6a03e7